Added Solutions to the homeworks
Please note that the rubric is available only for HW-1, HW-2, HW-3, HW-4, and HW-9 since sir did not allow the TAs to share it for every homework. Wherever the rubric is available, I have directly attached the rubric itself, and wherever it is not, I have attached a pdf containing my solutions. You would also find various codes along with the pdf, I added that so that it would be easy for you, and so that you could directly use the given code. There are also some codes having "correction" in their name. These are codes that are the correct solutions to those particular questions since all of my answers in the pdf are not correct. Please feel free to use these codes as a reference.